+++ docs/CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## Next
 
+* Inception V3: downloading weights, 
 * Tensors: added Load, Save for Local tensors.
+* Added HDF5 support.
 * Skip evaluation during test of demos.
 * Fixed dogsvscat demo's inconsistent mixed datasets issue, by yielding a correct spec.